+++ packages/template-retail-react-app/app/components/apple-pay-express/index.jsx
@@ -13,9 +13,8 @@ import {getCurrencyValueForApi} from '@salesforce/retail-react-app/app/component
 import {AdyenShippingMethodsService} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/shipping-methods'
 import {AdyenShippingAddressService} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/shipping-address'
 import {AdyenPaymentsService} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/payments'
-import {createTemporaryBasket, deleteTemporaryBasket} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/temporary-basket'
+import {createTemporaryBasket, deleteTemporaryBasket, cleanupTemporaryBasket} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/temporary-basket'
 import {getBasketWithTotals, forceOrderCalculation} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/basket-calculation'
-import {cleanupTemporaryBasket} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/basket-cleanup'
 import {useStandalonePaymentMethods} from '@salesforce/retail-react-app/app/components/apple-pay-express/hooks/use-standalone-payment-methods'
 import {useAccessToken} from '@salesforce/commerce-sdk-react'
 import useMultiSite from '@salesforce/retail-react-app/app/hooks/use-multi-site'
